Remove unused _data/*.yml files
In 87a397e8acba21a9a0dce6556740dc4132c182d4 the _data/devices directory was populated with .yml files that were essentially copies of the _data/*.yml files at the time and the Jekyll templates were updated to use the _data/devices/*.yml files.

This commit removes the now-unused _data/*.yml files in the hopes that it'll reduce confusion for future contributors to the project.

the verify.rb script is also updated so that the correct providers.yml file is parsed.